Throw exceptions instead of returning empty strings
When returning an empty string, Composer fails later in the installation process and returns a misleading error message:

> Failed to extract junaidbhura/polylang-pro: (9) '/usr/bin/unzip' -qq 'vendor/composer/tmp-XXXXXXXX' -d 'vendor/composer/ZZZZZZ'
>
> [vendor/composer/tmp-XXXXXXXX]
>   End-of-central-directory signature not found.  Either this file is not
>   a zipfile, or it constitutes one disk of a multi-part archive.  In the
>   latter case the central directory and zipfile comment will be found on
>   the last disk(s) of this archive.
> unzip:  cannot find zipfile directory in one of vendor/composer/tmp-XXXXXXXX or
>         vendor/composer/tmp-XXXXXXXX.zip, and cannot find vendor/composer/tmp-XXXXXXXX.ZIP, period.

This error stems from attempting to unzip nothing.

Throwing an exception during the processing of an add-on and of the download URL aborts the installation process sooner and provides a clearer reason for what went wrong.